+++ src/mesido/esdl/esdl_mixin.py
@@ -2,6 +2,7 @@ import base64
 import copy
 import dataclasses
 import logging
+import os
 import sys
 from datetime import timedelta
 from pathlib import Path
@@ -15,6 +16,7 @@ from mesido.component_type_mixin import (
 from mesido.esdl.asset_to_component_base import _AssetToComponentBase
 from mesido.esdl.common import Asset
 from mesido.esdl.edr_pipe_class import EDRGasPipeClass, EDRPipeClass
+from mesido.esdl.esdl_additional_vars_mixin import get_asset_contraints
 from mesido.esdl.esdl_heat_model import ESDLHeatModel
 from mesido.esdl.esdl_model_base import _ESDLModelBase
 from mesido.esdl.esdl_parser import ESDLStringParser
@@ -71,11 +73,15 @@ class ESDLMixin(
     for example demand profiles.
     """
 
+    csv_ensemble_mode: bool = False
+
     esdl_run_info_path: Path = None
 
     esdl_pi_validate_timeseries: bool = False
 
     __max_supply_temperature: Optional[float] = None
+
+    __esdl_ranged_constraint_usage: bool = False
 
     # TODO: remove this once ESDL allows specifying a minimum pipe size for an optional pipe.
     __minimum_pipe_size_name: str = "DN150"
@@ -101,6 +107,8 @@ class ESDLMixin(
         args : none
         kwargs : esdl_string or esdl_file_name must be provided
         """
+
+        self.__use_esdl_ranged_constraint: bool = kwargs.get("use_esdl_ranged_constraint", False)
 
         self.esdl_parser_class: type = kwargs.get("esdl_parser", ESDLStringParser)
         esdl_string = kwargs.get("esdl_string", None)
@@ -180,11 +188,13 @@ class ESDLMixin(
                 energy_system=self.__energy_system_handler.energy_system,
                 file_path=input_file_path,
                 database_credentials=read_only_dbase_credentials,
+                use_esdl_ranged_contraint=self._ESDLMixin__use_esdl_ranged_constraint,
             )
         else:  # read from a file, no database credentials needed
             self.__profile_reader: BaseProfileReader = profile_reader_class(
                 energy_system=self.__energy_system_handler.energy_system,
                 file_path=input_file_path,
+                use_esdl_ranged_contraint=self._ESDLMixin__use_esdl_ranged_constraint,
             )
 
         # This way we allow users to adjust the parsed ESDL assets
@@ -195,7 +205,13 @@ class ESDLMixin(
         name_to_id_map = {a.name: a.id for a in assets.values()}
 
         if isinstance(self, PhysicsMixin):
-            self.__model = ESDLHeatModel(assets, name_to_id_map, **self.esdl_heat_model_options())
+            self.__model = ESDLHeatModel(
+                assets=assets,
+                name_to_id_map=name_to_id_map,
+                esdl_version=self._ESDLMixin__energy_system_handler.energy_system.esdlVersion,
+                use_esdl_ranged_constraint=self._ESDLMixin__use_esdl_ranged_constraint,
+                **self.esdl_heat_model_options(),
+            )
         else:
             assert isinstance(self, QTHMixin)
 
@@ -265,6 +281,74 @@ class ESDLMixin(
                 return ii
         return None
 
+    def _get_pipe_max_size_input(
+        self,
+        asset: Asset,
+    ) -> str:
+        """
+        This function gets the upper limit for the pipe DN. Either from the asset attribute input
+        DN size or from the PipeDiameterConstraint, depending on what is applicable.
+        Args:
+            asset: mesido common asset with all attributes
+            max_size_attribute: type of attribute e.g. powrr, volume etc.
+            max_value_attribute: value that of the attribute,
+            constraint_attribute: Does the atrribute value originate from a constraint
+        """
+        # Backward compatibility:
+        # PipeDiameterConstraint vs pipe diameter attribute, as the value to be used for the
+        # maximum size of the pipe. We check the esdl version to determine which attribute to use
+        # for the maximum size of the pipe.
+        # While the front end is being updated to accomodate easy use of PipeDiameterConstraint:
+        # - If the esdl version caters for PipeDiameterConstraint:
+        #   - check if a constraint exists, then use it.
+        #   - if the contraint does not exist then use the diameter specific in the asset attribute
+        # - If PipeDiameterConstraint does not
+        esdl_version = self._ESDLMixin__energy_system_handler.energy_system.esdlVersion
+        # "v2602" contains the items needed for the ranged constraint implementation in MESIDO
+        if esdl_version is not None and esdl_version >= "v2602":
+            pipe_constraint, qty_pipe_constraint = get_asset_contraints(
+                self, asset, esdl.PipeDiameterConstraint
+            )
+
+            if qty_pipe_constraint > 1:
+                logger.exit(
+                    f"More than 1 pipe diameter constraint has been specified to "
+                    f"pipe named {asset.name}, currenlty only the 1st constraint is being used"
+                )
+                exit(1)
+            elif (
+                qty_pipe_constraint == 0
+                or pipe_constraint[0].maximum == esdl.PipeDiameterEnum.VALUE_SPECIFIED
+            ):
+                logger.warning(
+                    "Expected a pipe diameter contraint (upper size limit) for pipe named "
+                    f"{asset.name}, but none has been specified. Therefore, the pipe diameter "
+                    "specified in the asset attribute is being used."
+                )
+                # Do not delete the code below. This will be enforced late, once the front end has
+                # been updated for this
+
+                # get_potential_errors().add_potential_issue(
+                #     MesidoAssetIssueType.ASSET_UPPER_LIMIT,
+                #     asset.id,
+                #     f"Pipe named {asset.name}: The upper limit of the pipe size has to be
+                # specified via a maximum value in a pipe diameter constraint."
+                # )
+                # Raise the potential error here if applicable, with feedback to user
+                # Else a normal error exit might occer which will not give feedback to the user
+                # potential_error_to_error(self._error_type_check)
+
+                return asset.attributes["diameter"].name
+            else:
+                logger.warning(
+                    f"For pipe named {asset.name}, the pipe diameter constraint max value is "
+                    "used for the pipe's diameter upper limit, instead of the pipe diameter "
+                    "specified (if any) in the asset attribute."
+                )
+                return pipe_constraint[0].maximum.name
+        else:
+            return asset.attributes["diameter"].name
+
     def __override_pipe_classes_dicts(
         self,
         asset: Asset,
@@ -282,7 +366,7 @@ class ESDLMixin(
             c = override_classes[p] = []
             c.append(no_pipe_class)
 
-            max_size = asset.attributes["diameter"].name
+            max_size = self._get_pipe_max_size_input(asset)
 
             max_size_idx = [idx for idx, pipe in enumerate(pipe_classes) if pipe.name == max_size]
             assert len(max_size_idx) == 1
@@ -716,6 +800,19 @@ class ESDLMixin(
         None
         """
         super().read()
+        ensemble_size = 1
+        self.__ensemble = None
+        if self.csv_ensemble_mode:
+            self.__ensemble = np.genfromtxt(
+                os.path.join(self._input_folder, "ensemble.csv"),
+                delimiter=",",
+                deletechars="",
+                dtype=None,
+                names=True,
+                encoding=None,
+            )
+            ensemble_size = len(self.__ensemble)
+
         energy_system_components = self.energy_system_components
         esdl_carriers = self.esdl_carriers
         self.hot_cold_pipe_relations()
@@ -726,7 +823,8 @@ class ESDLMixin(
             esdl_asset_id_to_name_map=self.esdl_asset_id_to_name_map,
             esdl_assets=self.esdl_assets,
             carrier_properties=esdl_carriers,
-            ensemble_size=self.ensemble_size,
+            ensemble_size=ensemble_size,
+            ensemble=self.__ensemble,
         )
 
     def write(self) -> None:
@@ -851,3 +949,9 @@ class ESDLMixin(
                 filtered_assets[asset_id] = asset_type
 
         return filtered_assets
+
+    def ensemble_member_probability(self, ensemble_member):
+        if self.csv_ensemble_mode:
+            return self.__ensemble["probability"][ensemble_member]
+        else:
+            return 1.0